aider: Update urllib3 dependency version constraint to allow any version up to `2.0.0` in the `pyproject.toml` file.
# Aider chat conversation:

USER: Is there a way to get poetry to determine the current version and update the pyproject file?

ASSISTANT: Poetry does not automatically determine the current version of your project from external sources like Git tags. However, you can manually update the version in the `pyproject.toml` file using the `poetry version` command. This command updates the version in the `pyproject.toml` file and can also create a corresponding Git tag if you use it with the `-s` or `--sync` option.

Here's how you can use the `poetry version` command:

1. To increment the version interactively, just run:
   ```sh
   poetry version
   ```
   This will prompt you to select the part of the version to update (patch, minor, major, etc.).

2. To set a specific version, use:
   ```sh
   poetry version <new_version>
   ```
   Replace `<new_version>` with the desired version number.

3. To increment the version and create a Git tag, use:
   ```sh
   poetry version <new_version> -s
   ```

Please note that the `poetry version` command does not automatically determine the next version number; you must specify it or choose the increment (patch, minor, major).

If you want to automate the versioning based on Git tags or commits, you would need to use external tools or scripts in conjunction with Poetry. For example, you could use `git describe` to get the latest tag and then use this information to set the version with `poetry version`.

If you have a specific version number in mind that you want to set in the `pyproject.toml` file, please provide it, and I can generate the diff for you.

USER: Is there a way to get poetry to determine the current version of a depency and update the pyproject file for that dependency?

ASSISTANT: Poetry does not have a built-in command to automatically update the `pyproject.toml` file with the latest version of a specific dependency. However, you can update dependencies to their latest versions by using the `poetry update` command. This command will update the `pyproject.toml` file and the lock file with the latest versions of the dependencies that match the version constraints specified.

If you want to update a specific dependency to its latest version within the constraints defined in `pyproject.toml`, you can run:

```sh
poetry update <dependency-name>
```

Replace `<dependency-name>` with the name of the dependency you want to update.

If you want to change the version constraints for a dependency to allow for newer versions, you would manually edit the `pyproject.toml` file to adjust the version range and then run `poetry update` to apply the changes.

For example, if you want to change the version constraint for `urllib3` to allow any version up to `2.0.0` (excluding `2.0.0`), you would modify the `pyproject.toml` file like this:

```diff
--- pyproject.toml
+++ pyproject.toml
@@ ... @@
 [tool.poetry.dependencies]
 python = "^3.8"
-urllib3 = ">= 1.25.3, < 2.1.0"
+urllib3 = ">= 1.25.3, < 2.0.0"
 python-dateutil = ">= 2.5.3"
 pydantic = "^1.10.5, <2"
 aenum = ">=3.1.11"
```

After making this change, you would run `poetry update urllib3` to update `urllib3` to the latest version that satisfies the new constraint.

If you have a specific dependency and version constraint in mind, please provide the details, and I can generate the diff for the `pyproject.toml` file.