aider: Updated the `generate-api` script to allow `.openapi-generator*` config files to be in a subdirectory.
# Aider chat conversation:

USER: Can I run openapi-generator-cli in such a way the .openapi* config files are in a subdirectory?

ASSISTANT: Yes, you can run `openapi-generator-cli` with configuration files located in a subdirectory. You would typically use the `-c` or `--config` option to specify the path to your configuration file when running the generator. Here's an example of how you might invoke the generator with a configuration file in a subdirectory:

```bash
openapi-generator-cli generate -i your-openapi-file.yaml -g your-generator-name -c path/to/config/.openapirc.yaml
```

In this command:

- `-i` specifies the input OpenAPI definition file.
- `-g` specifies the generator name (e.g., `java`, `python`, etc.).
- `-c` specifies the path to the configuration file, which can be in any directory.

If you have multiple configuration files or you want to organize your configuration files in a specific directory, you can point to the appropriate file within that directory using the `-c` option.

USER: I added these *read-write* files: /home/harleypig/projects/bw-serve-client/bin/generate-api

ASSISTANT: To modify the `generate-api` script to allow the `.openapi-generator*` config files to be in a subdirectory, you can update the `--config` option to point to the new location of your configuration file. Let's assume you want to place your configuration file in a subdirectory called `config`.

Here is the diff for the changes:

```diff
--- /home/harleypig/projects/bw-serve-client/bin/generate-api
+++ /home/harleypig/projects/bw-serve-client/bin/generate-api
@@ ... @@
-config_file='.generator-pydantic.yml'
+config_file='config/.generator-pydantic.yml'
@@ ... @@
-genopts+=('--config' "$config_file")
+genopts+=('--config' "$config_file")
```

This change updates the `config_file` variable to point to the new location within the `config` subdirectory. The `--config` option in the `genopts` array will now correctly reference the configuration file in its new location.