aider: Imported the settings from `tox.ini` into `pyproject.toml`.
# Aider chat conversation:

USER: Import the settings in tox.ini into pyproject.toml

ASSISTANT: To import the settings from `tox.ini` into `pyproject.toml`, we will add a `[tool.tox]` section to `pyproject.toml` and replicate the settings from `tox.ini`. Here is the diff for the changes:

```diff
--- /home/harleypig/projects/bw-serve-client/pyproject.toml
+++ /home/harleypig/projects/bw-serve-client/pyproject.toml
@@ ... @@
 [tool.poetry.group.docs.dependencies]
 mkdocs = "^1.5.3"
 sphinx = "^7.2.6"
 sphinx-autodoc-typehints = "^2.0.0"
 sphinxcontrib-email = "^0.3.6"
 sphinxcontrib-openapi = "^0.8.4"
 sphinx-lint = "^0.9.1"
+
+[tool.tox]
+envlist = "py3"
+
+[testenv]
+deps = [
+    "-r{toxinidir}/requirements.txt",
+    "-r{toxinidir}/test-requirements.txt"
+]
+commands = [
+    "pytest --cov=bw_serve_client"
+]
```

This diff adds a new `[tool.tox]` section to `pyproject.toml` with the same settings as found in `tox.ini`. Please note that the actual syntax for integrating Tox settings into `pyproject.toml` may vary depending on the version of Tox and any plugins you are using. The above is a representation of how it might look. If Tox or any related tooling in your project requires a different format, please adjust accordingly.