aider: Add a mypy configuration section specifying the HTML report directory in `.openapi/mypy-reports` in the `pyproject.toml` file.
# Aider chat conversation:

USER: I want to change the value the mustache variable appName is using. How do I do that?

ASSISTANT: To change the value that the mustache variable `appName` is using, you would typically need to modify the configuration or input file that provides the value for this variable when the mustache templates are rendered. This is often done in a configuration file or script that invokes the `openapi-generator-cli` or similar tool that uses these templates.

If you have a specific new value in mind for the `appName` variable, please provide it, and I can guide you on where to make the change. If the change is to be made in a file that is not currently read-write, you will need to add it to the chat so I can assist you with the necessary modifications.

USER: I added these *read-write* files: /home/harleypig/projects/bw-serve-client/pyproject.toml

ASSISTANT: To change the value of the `appName` variable in the `pyproject.toml` file, you would typically update the `[tool.poetry]` section where the `name` attribute is defined. This attribute often corresponds to the `appName` variable used in mustache templates.

Please provide the new value you want to set for `appName`, and I will create the diff for you. If you're not sure where the `appName` is coming from or if it's not from the `name` attribute in `pyproject.toml`, please provide additional context or instructions.

USER: Add a mypy configuration section. Since we're starting out I just want to make sure what we have is working correctly, let's start out with a minimal configuration. Have it create an html report in .openapi/mypy-reports.

ASSISTANT: To add a minimal mypy configuration section to the `pyproject.toml` file with an HTML report output to `.openapi/mypy-reports`, you can add the following `[tool.mypy]` section:

```diff
--- /home/harleypig/projects/bw-serve-client/pyproject.toml
+++ /home/harleypig/projects/bw-serve-client/pyproject.toml
@@ ... @@
 [tool.poetry.group.docs.dependencies]
 mkdocs = "^1.5.3"
 sphinx = "^7.2.6"
 sphinx-autodoc-typehints = "^2.0.0"
 sphinxcontrib-email = "^0.3.6"
 sphinxcontrib-openapi = "^0.8.4"
 sphinx-lint = "^0.9.1"
+
+[tool.mypy]
+html_report = ".openapi/mypy-reports"
```

This diff adds a new `[tool.mypy]` section to the `pyproject.toml` file, specifying the directory for the HTML report generated by mypy. This is a minimal configuration, and you can add more options to the `[tool.mypy]` section as needed to customize mypy's behavior further.