aider: Created a logging class in `bw_serve_client/logging.py` to allow users to pass in their own logging function.
# Aider chat conversation:

USER: Create a logging class in bw_serve_client/logging.py that uses pythons logging module that will be used by other classes in this library. Write it in such a way that users of the library can pass in their own logging function, assuming the same interface is available (e.g., logger.debug, logger.info, etc.).

ASSISTANT: To proceed with creating the logging class in `bw_serve_client/logging.py`, I need you to add this file to the chat as a *read-write* file. Please do so to enable modifications.

USER: It needs to be created.

ASSISTANT: Here is the diff for creating the new `bw_serve_client/logging.py` file with a logging class that uses Python's logging module. This class allows users of the library to pass in their own logging function.

```diff
--- /dev/null
+++ bw_serve_client/logging.py
@@ ... @@
+import logging
+
+class LibraryLogger:
+    def __init__(self, logger=None):
+        if logger is None:
+            self.logger = logging.getLogger(__name__)
+            handler = logging.StreamHandler()
+            formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+            handler.setFormatter(formatter)
+            self.logger.addHandler(handler)
+            self.logger.setLevel(logging.DEBUG)
+        else:
+            self.logger = logger
+
+    def debug(self, msg, *args, **kwargs):
+        self.logger.debug(msg, *args, **kwargs)
+
+    def info(self, msg, *args, **kwargs):
+        self.logger.info(msg, *args, **kwargs)
+
+    def warning(self, msg, *args, **kwargs):
+        self.logger.warning(msg, *args, **kwargs)
+
+    def error(self, msg, *args, **kwargs):
+        self.logger.error(msg, *args, **kwargs)
+
+    def critical(self, msg, *args, **kwargs):
+        self.logger.critical(msg, *args, **kwargs)
```

This class, `LibraryLogger`, can be used directly or with a custom logger passed as an argument. If no custom logger is provided, it sets up a basic logger with a stream handler and a simple formatter.