aider: Add both FileHandler and StreamHandler to BSCLogger for dual logging capabilities.
# Aider chat conversation:

USER: Can both FileHandler and StreamHandler be set?

ASSISTANT: Yes, both `FileHandler` and `StreamHandler` can be set up in the same logger to output logs to both a file and the console. Here's how you can modify the `BSCLogger` class in `bw_serve_client/logging.py` to include both handlers:

```diff
--- bw_serve_client/logging.py
+++ bw_serve_client/logging.py
@@ ... @@
 class BSCLogger:
     def __init__(self, logger=None):
         if logger is None:
             self.logger = logging.getLogger('bw-serve-client')
-            handler = logging.FileHandler('bw-serve-client.log')
+            file_handler = logging.FileHandler('bw-serve-client.log')
+            stream_handler = logging.StreamHandler()
             formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-            handler.setFormatter(formatter)
-            self.logger.addHandler(handler)
+            file_handler.setFormatter(formatter)
+            stream_handler.setFormatter(formatter)
+            self.logger.addHandler(file_handler)
+            self.logger.addHandler(stream_handler)
             self.logger.setLevel(logging.DEBUG)
         else:
             self.logger = logger
```

This modification adds a `StreamHandler` alongside the existing `FileHandler`. Both handlers use the same formatter and are added to the logger. This setup allows logging output to be directed both to the console and to the file `bw-serve-client.log`.